Move OasAgent::Control to correct place
